resolvido o problema da opcional
