+++ babySteps/babySteps/Views/Login/LoginProfileView.swift
@@ -11,6 +11,7 @@ import Combine
 
 
 struct LoginProfileView: View {
+    
     @State private var stackPath = Router()
     @State private var viewModel = LoginProfileViewModel()
     
@@ -18,8 +19,120 @@ struct LoginProfileView: View {
 
         VStack{
             if viewModel.isLoggedIn {
-                LoggedProfileView()
-                //                LoggedProfileView(teacher: viewModel.teacher)
+                //LoggedProfileView(viewModel: viewModel)
+                
+                VStack {
+                    VStack(spacing: 15) {
+
+                        AsyncImage(url: URL(string: "")) { image in
+                            image
+                                .resizable()
+                                .clipShape(Circle())
+                                .frame(width: 112, height: 112)
+                        } placeholder: {
+                            Image("childPlaceHolder")
+                                .resizable()
+                                .clipShape(Circle())
+                                .frame(width: 112, height: 112)
+                        }
+                        
+                        Text(viewModel.teacherName)
+                            .font(.system(.title2, design: .rounded))
+                            //                            .fontWeight(.regular)
+//                        Text("\(viewModel.teacher?.firstName ?? "Nao há first Name") \(viewModel.teacher?.lastName ?? "")")
+//                            .font(.system(.title, design: .rounded))
+//                            .bold()
+
+//                        Text(viewModel.teacher?.email ?? "")
+//                            .font(.system(.title2, design: .rounded))
+//                            .fontWeight(.regular)
+                    }
+
+                    VStack(alignment: .leading) {
+                        Section {
+                            ClassProfileCardView()
+                        } header: {
+                            Text("Minhas Turmas")
+                                .font(.system(.title, design: .rounded))
+                                .bold()
+                            
+                            //listar aqui as turmas do professor
+                        }
+                    }
+                    .frame(maxWidth: .infinity, alignment: .topLeading)
+                    HStack {
+                        VStack(alignment: .leading) {
+                            Section {
+                                CalendarProfileCardView()
+
+                            } header: {
+                                Text("Agenda")
+                                    .font(.system(.title, design: .rounded))
+                                    .bold()
+                            }
+                        }
+                        .frame(maxWidth: .infinity, alignment: .topLeading)
+
+                        VStack(alignment: .leading) {
+                            Section {
+                                
+                                Text("lista aqui")
+        //                        List {
+        //                            ForEach(viewModel.teacher!.announcements, id: \.self) { announcement in
+        //                                Text(announcement.title)
+        //                                .font(.system(.body, design: .rounded, weight: .regular))
+        //                                                        }
+        //                        }.frame(width: .infinity)
+        //                        .listRowSpacing(10)
+
+                            } header: {
+                                HStack(alignment: .bottom) {
+                                    Text("Lembretes")
+                                        .font(.system(.title, design: .rounded))
+                                        .bold()
+                                    Spacer()
+                                    Button {
+                                        print("EditPressed")
+                                    } label: {
+                                        Text("Editar")
+                                            .font(.system(.body, design: .rounded, weight: .regular))
+                                            .foregroundStyle(Color.orangePrimary)
+                                    }
+                                }
+                                
+                            } footer: {
+                                Button {
+                                    print("AddPressed")
+                                } label: {
+                                    HStack {
+                                        Image(systemName: "plus")
+                                        Text("Adicionar novo lembrete")
+                                            .font(.system(.body, design: .rounded, weight: .regular))
+                                    }.foregroundStyle(Color.orangePrimary)
+                                        .background(Color.clear)
+                                }
+                            }
+                        }
+                    }
+                    Button(action: {
+                        print("EndSession")
+                        //inserir lógica de deslogar
+                    }, label: {
+                        Text("Finalizar Sessão")
+                            .padding(8)
+                            .font(.system(size: 19, weight: .regular, design: .rounded))
+                            .foregroundStyle(Color.orangePrimary)
+                            .frame(width: 500)
+                            .background(Color.white)
+                            .clipShape(.rect(cornerRadius: 12))
+                    
+                    })
+
+                }
+                .padding()
+                    .frame(maxWidth: /*@START_MENU_TOKEN@*/.infinity/*@END_MENU_TOKEN@*/, maxHeight: .infinity)
+                    .background(Color.detailViewBackground)
+                
                 } else {
                 Text("Tela de Login")
                 
@@ -40,6 +153,43 @@ struct LoginProfileView: View {
     
 }
 
+
+struct ClassProfileCardView: View {
+    var body: some View {
+        VStack(alignment: .leading, spacing: 0) {
+            
+            Text("Turma B")
+                .padding(5)
+                .font(.system(.title, design: .rounded))
+                .bold()
+                .background(Color.white)
+                .clipShape(.rect(cornerRadii: RectangleCornerRadii(topLeading: 16, topTrailing: 15)))
+            
+            VStack(alignment: .leading) {
+                Text("Número de alunos: 10")
+                    .font(.system(.body, design: .rounded))
+                Text("Período: Integral")
+                    .font(.system(.body, design: .rounded))
+            }
+            .padding(10)
+            .frame(maxWidth: .infinity, alignment: .bottomLeading)
+            .background(Color.white)
+            
+        }
+        .frame(width: 280, height: 150, alignment: .bottomLeading)
+        
+        .background {
+            AsyncImage(url: URL(string: "")) { image in
+                image
+            } placeholder: {
+                Image("classImagePlaceHolder")
+                    .resizable()
+                    .scaledToFill()
+            }
+        }
+        .clipShape(.rect(cornerRadius: 16))
+    }
+}
 
 #Preview {
     LoginProfileView()
